requirements.txt
# pip install -r requirements.txt
plexapi
requests
trakt
ruamel.yaml
lxml
